requirements.txt
@"
streamlit
pandas
numpy
plotly
google-api-python-client
google-generativeai
beautifulsoup4
requests
"@ | Out-File -FilePath requirements.txt -Encoding UTF8
